+++ cli/Sources/tuist/TuistCommand.swift
@@ -132,6 +132,7 @@ public struct TuistCommand: AsyncParsableCommand {
         #endif
     }
 
+    // swiftlint:disable:next function_body_length
     public static func main(
         logFilePath: AbsolutePath,
         _ arguments: [String]? = nil,
@@ -226,21 +227,23 @@ public struct TuistCommand: AsyncParsableCommand {
             }
         #else
             try await withLoggerForNoora(logFilePath: logFilePath) {
-                try await Noora.$current.withValue(initNoora()) {
-                    do {
-                        var command = try parseAsRoot(processedArguments)
+                do {
+                    let command = try parseAsRoot(processedArguments)
+                    let jsonThroughNoora = (command as? NooraReadyCommand)?.jsonThroughNoora ?? false
+                    try await Noora.$current.withValue(initNoora(jsonThroughNoora: jsonThroughNoora)) {
                         if var asyncCommand = command as? AsyncParsableCommand {
                             try await asyncCommand.run()
                         } else {
-                            try command.run()
+                            var mutableCommand = command
+                            try mutableCommand.run()
                         }
                         outputCompletion(
                             logFilePath: logFilePath,
                             shouldOutputLogFilePath: false
                         )
-                    } catch {
-                        onError(error, isParsingError: false, logFilePath: logFilePath)
                     }
+                } catch {
+                    onError(error, isParsingError: false, logFilePath: logFilePath)
                 }
             }
         #endif